+++ ext/function.c
@@ -52,7 +52,7 @@ static VALUE rb_gsl_function_set_f(int argc, VALUE *argv, VALUE obj)
     CHECK_PROC(argv[0]);
     rb_ary_store(ary, 0, argv[0]);
     ary2 = rb_ary_new2(argc-1);
-    for (i = 1; i < argc; i++) rb_ary_store(ary2, i-1, argv[i]);
+    for (i = 1; (int) i < argc; i++) rb_ary_store(ary2, i-1, argv[i]);
     rb_ary_store(ary, 1, ary2);
     break;
   }
@@ -221,7 +221,7 @@ static VALUE rb_gsl_function_set_params(int argc, VALUE *argv, VALUE obj)
     rb_ary_store(ary, 1, argv[0]);
   } else {
     ary2 = rb_ary_new2(argc);
-    for (i = 0; i < argc; i++) rb_ary_store(ary2, i, argv[i]);
+    for (i = 0; (int) i < argc; i++) rb_ary_store(ary2, i, argv[i]);
     rb_ary_store(ary, 1, ary2);
   }
   return obj;
@@ -337,7 +337,7 @@ static VALUE rb_gsl_function_fdf_new(int argc, VALUE *argv, VALUE klass)
   F->params = (void *) ary;
   rb_ary_store(ary, 2, Qnil);
   rb_ary_store(ary, 3, Qnil);
-  for (i = 0; i < argc; i++) setfunc(i, argv, F);
+  for (i = 0; (int) i < argc; i++) setfunc(i, argv, F);
   return Data_Wrap_Struct(klass, gsl_function_fdf_mark, gsl_function_fdf_free, F);
 }
 
@@ -360,7 +360,7 @@ static VALUE rb_gsl_function_fdf_set(int argc, VALUE *argv, VALUE obj)
   ary = (VALUE) F->params;
   rb_ary_store(ary, 2, Qnil);
   rb_ary_store(ary, 3, Qnil);
-  for (i = 0; i < argc; i++) setfunc(i, argv, F);
+  for (i = 0; (int) i < argc; i++) setfunc(i, argv, F);
   return obj;
 }
 
@@ -427,7 +427,7 @@ static VALUE rb_gsl_function_fdf_set_params(int argc, VALUE *argv, VALUE obj)
     rb_ary_store(ary, 3, argv[0]);
   } else {
     ary2 = rb_ary_new2(argc);
-    for (i = 0; i < argc; i++) rb_ary_store(ary2, i, argv[i]);
+    for (i = 0; (int) i < argc; i++) rb_ary_store(ary2, i, argv[i]);
     rb_ary_store(ary, 3, ary2);
   }
   return obj;
