+++ ext/poly2.c
@@ -37,7 +37,7 @@ static gsl_poly_int* mygsl_poly_hermite(int n1)
     p2 = gsl_vector_int_calloc(n1 + 1);
     memcpy(p1->data, coef2, 3*sizeof(int));
     memcpy(p2->data, coef1, 2*sizeof(int));
-    for (n = 2; n < n1; n++) {
+    for (n = 2; (int) n < n1; n++) {
       gsl_vector_int_memcpy(p0, p1);
       mygsl_vector_int_shift_scale2(p0, n);
       gsl_vector_int_scale(p2, 2*n);
@@ -77,7 +77,7 @@ static gsl_poly_int* mygsl_poly_cheb(int n1)
     p2 = gsl_vector_int_calloc(n1 + 1);
     memcpy(p1->data, coef2, 3*sizeof(int));
     memcpy(p2->data, coef1, 2*sizeof(int));
-    for (n = 2; n < n1; n++) {
+    for (n = 2; (int) n < n1; n++) {
       gsl_vector_int_memcpy(p0, p1);
       mygsl_vector_int_shift_scale2(p0, n);
 
@@ -116,7 +116,7 @@ static gsl_poly_int* mygsl_poly_chebII(int n1)
     p2 = gsl_vector_int_calloc(n1 + 1);
     memcpy(p1->data, coef2, 3*sizeof(int));
     memcpy(p2->data, coef1, 2*sizeof(int));
-    for (n = 2; n < n1; n++) {
+    for (n = 2; (int) n < n1; n++) {
       gsl_vector_int_memcpy(p0, p1);
       mygsl_vector_int_shift_scale2(p0, n);
       gsl_vector_int_sub(p0, p2);
@@ -148,7 +148,7 @@ static gsl_poly_int* mygsl_poly_laguerre(int n)
     break;
   default:
     k = gsl_sf_fact(n);
-    for (m = 0; m <= n; m++) {
+    for (m = 0; (int) m <= n; m++) {
       val = k*k/gsl_sf_fact(n-m)/gsl_pow_2(gsl_sf_fact(m));
       if (m%2 == 1) val *= -1;
       gsl_vector_int_set(p0, m, val);
@@ -164,7 +164,7 @@ static gsl_poly_int* mygsl_poly_bessel(int n)
   gsl_vector_int *p0;
   if (n < 0) rb_raise(rb_eArgError, "order must be >= 0");
   p0 = gsl_vector_int_calloc(n + 1);
-  for (k = 0; k <= n; k++) {
+  for (k = 0; (int) k <= n; k++) {
     gsl_vector_int_set(p0, k, gsl_sf_fact(n+k)/gsl_sf_fact(n-k)/gsl_sf_fact(k)/((int) pow(2, k)));
   }
   return p0;
@@ -191,7 +191,7 @@ static gsl_poly_int* mygsl_poly_bell(int n1)
   default:
     p1 = gsl_vector_int_calloc(n1 + 1);
     memcpy(p1->data, coef2, 3*sizeof(int));
-    for (n = 2; n < n1; n++) {
+    for (n = 2; (int) n < n1; n++) {
       gsl_vector_int_memcpy(p0, p1);
       mygsl_vector_int_shift(p0, n);
       for (j = 0; j < n; j++) {
